# packages/mtg-core/src/mtg_core/api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MTG Core API",
    "description": "Magic: The Gathering card database and analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/user/decks": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "List Decks",
        "description": "List all decks.",
        "operationId": "list_decks_user_decks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeckSummaryResponse"
                  },
                  "type": "array",
                  "title": "Response List Decks User Decks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create Deck",
        "description": "Create a new deck.",
        "operationId": "create_deck_user_decks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDeckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/decks/{deck_id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Deck",
        "description": "Get a deck with all its cards, enriched with card data.",
        "operationId": "get_deck_user_decks__deck_id__get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update Deck",
        "description": "Update deck metadata.",
        "operationId": "update_deck_user_decks__deck_id__put",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Deck User Decks  Deck Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete Deck",
        "description": "Delete a deck.",
        "operationId": "delete_deck_user_decks__deck_id__delete",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDeckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/decks/{deck_id}/cards": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add Card To Deck",
        "description": "Add a card to a deck.",
        "operationId": "add_card_to_deck_user_decks__deck_id__cards_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/decks/{deck_id}/cards/{card_name}": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Remove Card From Deck",
        "description": "Remove a card from a deck.",
        "operationId": "remove_card_from_deck_user_decks__deck_id__cards__card_name__delete",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          },
          {
            "name": "card_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Name"
            }
          },
          {
            "name": "sideboard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sideboard"
            }
          },
          {
            "name": "maybeboard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Maybeboard"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update Card Quantity",
        "description": "Update card quantity in a deck.",
        "operationId": "update_card_quantity_user_decks__deck_id__cards__card_name__put",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          },
          {
            "name": "card_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Name"
            }
          },
          {
            "name": "sideboard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sideboard"
            }
          },
          {
            "name": "maybeboard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Maybeboard"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardQuantityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/decks/{deck_id}/cards/{card_name}/move": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Move Card In Deck",
        "description": "Move a card between mainboard, sideboard, and maybeboard.",
        "operationId": "move_card_in_deck_user_decks__deck_id__cards__card_name__move_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          },
          {
            "name": "card_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveCardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/decks/{deck_id}/analyze/mana-curve": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Analyze Deck Mana Curve By Id",
        "description": "Analyze deck mana curve by deck ID.",
        "operationId": "analyze_deck_mana_curve_by_id_user_decks__deck_id__analyze_mana_curve_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManaCurveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/decks/{deck_id}/analyze/colors": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Analyze Deck Colors By Id",
        "description": "Analyze deck colors by deck ID.",
        "operationId": "analyze_deck_colors_by_id_user_decks__deck_id__analyze_colors_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColorAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/decks/{deck_id}/analyze/price": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Analyze Deck Price By Id",
        "description": "Analyze deck price by deck ID.",
        "operationId": "analyze_deck_price_by_id_user_decks__deck_id__analyze_price_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddCardResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "AddCardResponse",
        "description": "Response after adding a card."
      },
...
      "CreateDeckRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "commander": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commander"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateDeckRequest",
        "description": "Request to create a new deck."
      },
      "CreateDeckResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateDeckResponse",
        "description": "Response after creating a deck."
      },
...
      "DeckSummaryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "card_count": {
            "type": "integer",
            "title": "Card Count"
          },
          "sideboard_count": {
            "type": "integer",
            "title": "Sideboard Count"
          },
          "maybeboard_count": {
            "type": "integer",
            "title": "Maybeboard Count"
          },
          "commander": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commander"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "format",
          "card_count",
          "sideboard_count",
          "maybeboard_count",
          "commander",
          "updated_at"
        ],
        "title": "DeckSummaryResponse",
        "description": "Summary of a deck for listing."
      },
...
      "DeleteDeckResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeleteDeckResponse",
        "description": "Response after deleting a deck."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDeckRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "commander": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commander"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "UpdateDeckRequest",
        "description": "Request to update deck metadata."
      },
...
    }
  }
}